BE: - Added fields for recipe and ingredients in domain class and DTO class, updated getters and setters. - Added getByID controller url with the corresponding method in service UI: - Added a posibility to review the information (currently only id and name) of the dish, by clicking on it (using ActivatedRoute).

+++ src/main/java/com/cookBook/dto/DishDTO.java
@@ -3,7 +3,7 @@ package com.cookBook.dto;
 import com.cookBook.entity.Dish;
 
 public class DishDTO {
-    private Integer id;
+    private Long id;
     private String name;
     private String ingredients;
     private String description;
@@ -16,11 +16,11 @@ public class DishDTO {
         return this;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
